- converted PackageDeployer & FileFetcher into .ts files - moved 'mkdir()' function into extenstions project since it can be defined as an extensions off of the 'fs' module
