--- src/fetchers/fetchers.module.js
@@ -1,6 +1,0 @@
-let FileFetcher = require('./fetchers.file');
-
-
-module.exports = {
-    FileFetcher: FileFetcher
-}
